working on 0.2.0; getting rid of pyopenssl
