+++ CHANGES.txt
@@ -1,9 +1,12 @@
 0.2.0
-    replacing pyopenssl with cryptography as much as possible
-    pyopenssl was originally used because certbot required it and this package
-    was originally developed to troubleshoot certbot installations;
-    however certbot uses cryptography, so we can consider that to be available
-    and it is the officially preferred interface by both projects and the psf 
+    * replaced pyopenssl with cryptography.
+      pyopenssl was originally used because certbot required it and this package
+      was originally developed to troubleshoot certbot installations;
+      however certbot uses cryptography, so we can consider that to be available
+      and it is the officially preferred interface by both projects and the psf.
+    * split the package into multiple files
+      this is for internal use and organization is subject to change.
+      please access the main namespace.
 
 0.1.9
     pulled ari serial_no decoder into a separate unction
