+++ setup.cfg
@@ -13,9 +13,15 @@ import_order_style = appnexus
 per-file-ignores =
 	setup.py: E501
     src/cert_utils/core.py: E501
+    src/cert_utils/compat/cryptography.py: E501
+    src/cert_utils/compat/openssl.py: E501
+    src/cert_utils/compat/pyopenssl.py: E501
+    src/cert_utils/conditionals.py: E501
+    src/cert_utils/convert.py: E501
     src/cert_utils/letsencrypt_info.py: E501
+    src/cert_utils/model.py: E501
+    src/cert_utils/scripts/cert_info.py: E501
     src/cert_utils/utils.py: E203, E501
-
     tests/_utils.py: E501    
     tests/test_data/long_chains/TestA/_demo.py: E501
     tests/test_unit.py: E501, F841
