+++ src/cert_utils/__init__.py
@@ -1,5 +1,7 @@
+from .convert import *  # noqa: F401, F403
 from .core import *  # noqa: F401, F403
+from .utils import *  # noqa: F401, F403
 
 # ==============================================================================
 
-__VERSION__ = "0.2.0"
+__VERSION__ = "0.2.0dev"
