+++ src/cert_utils/core.py
@@ -9,8 +9,6 @@ import os
 import re
 import subprocess
 import tempfile
-import textwrap
-from types import ModuleType
 from typing import Any
 from typing import Dict
 from typing import List
@@ -24,6 +22,30 @@ from dateutil import parser as dateutil_parser
 import psutil
 
 # localapp
+from .compat.cryptography import _cryptography__cert_or_req__all_names
+from .compat.cryptography import _cryptography__public_key_spki_sha256
+from .compat.cryptography import _cryptography__public_key_technology
+from .compat.cryptography import _format_cryptography_components
+from .compat.cryptography import cryptography__cert_and_chain_from_fullchain
+from .compat.openssl import _cleanup_openssl_modulus
+from .compat.openssl import _format_openssl_components
+from .compat.openssl import _openssl_cert__normalize_pem
+from .compat.openssl import _openssl_cert_single_op__pem_filepath
+from .compat.openssl import _openssl_spki_hash_cert
+from .compat.openssl import _openssl_spki_hash_csr
+from .compat.openssl import _openssl_spki_hash_pkey
+from .compat.openssl import cert_ext__pem_filepath
+from .compat.openssl import csr_single_op__pem_filepath
+from .compat.openssl import key_single_op__pem_filepath
+from .conditionals import crypto_dsa  # Optional[pypi:cryptography]
+from .conditionals import crypto_ec  # Optional[pypi:cryptography]
+from .conditionals import crypto_ed25519  # Optional[pypi:cryptography]
+from .conditionals import crypto_ed448  # Optional[pypi:cryptography]
+from .conditionals import crypto_hashes  # Optional[pypi:cryptography]
+from .conditionals import crypto_rsa  # Optional[pypi:cryptography]
+from .conditionals import crypto_serialization  # Optional[pypi:cryptography]
+from .conditionals import cryptography  # Optional[pypi:cryptography]
+from .conditionals import josepy  # Optional[pypi:josepy]
 from .errors import CryptographyError
 from .errors import FallbackError_FilepathRequired
 from .errors import OpenSslError
@@ -33,8 +55,20 @@ from .errors import OpenSslError_InvalidCSR
 from .errors import OpenSslError_InvalidKey
 from .errors import OpenSslError_VersionTooLow
 from .model import KeyTechnology
+from .utils import _cert_pubkey_technology__text
+from .utils import _csr_pubkey_technology__text
+from .utils import authority_key_identifier_from_text
+from .utils import cleanup_pem_text
 from .utils import convert_binary_to_hex
-from .utils import cryptography__cert_and_chain_from_fullchain
+from .utils import issuer_uri_from_text
+from .utils import jose_b64
+from .utils import new_pem_tempfile
+from .utils import RE_openssl_x509_subject
+from .utils import san_domains_from_text
+from .utils import serial_from_text
+from .utils import split_pem_chain
+
+# from .utils import new_der_tempfile
 
 # ==============================================================================
 
@@ -42,7 +76,6 @@ if TYPE_CHECKING:
     import datetime
     from cryptography.x509 import Certificate
     from cryptography.x509 import CertificateSigningRequest
-    from cryptography.x509.name import Name
     from cryptography.hazmat.primitives.hashes import HashAlgorithm
     from cryptography.hazmat.primitives.asymmetric.dsa import DSAPrivateKey
     from cryptography.hazmat.primitives.asymmetric.dsa import DSAPublicKey
@@ -75,51 +108,6 @@ if TYPE_CHECKING:
         X448PublicKey,
     ]
 
-# ------------------------------------------------------------------------------
-# Conditional Imports
-
-cryptography: Optional[ModuleType]
-crypto_dsa: Optional[ModuleType]
-crypto_ec: Optional[ModuleType]
-crypto_ed25519: Optional[ModuleType]
-crypto_ed448: Optional[ModuleType]
-crypto_hashes: Optional[ModuleType]
-crypto_padding: Optional[ModuleType]
-crypto_pkcs7: Optional[ModuleType]
-crypto_rsa: Optional[ModuleType]
-crypto_serialization: Optional[ModuleType]
-crypto_x509: Optional[ModuleType]
-josepy: Optional[ModuleType]
-
-try:
-    import cryptography
-    import cryptography.x509 as crypto_x509
-    from cryptography.hazmat.primitives import hashes as crypto_hashes
-    from cryptography.hazmat.primitives import padding as crypto_padding
-    from cryptography.hazmat.primitives import serialization as crypto_serialization
-    from cryptography.hazmat.primitives.asymmetric import dsa as crypto_dsa
-    from cryptography.hazmat.primitives.asymmetric import ec as crypto_ec
-    from cryptography.hazmat.primitives.asymmetric import ed25519 as crypto_ed25519
-    from cryptography.hazmat.primitives.asymmetric import ed448 as crypto_ed448
-    from cryptography.hazmat.primitives.asymmetric import rsa as crypto_rsa
-    from cryptography.hazmat.primitives.serialization import pkcs7 as crypto_pkcs7
-except ImportError:
-    cryptography = None
-    crypto_x509 = None
-    crypto_dsa = None
-    crypto_ec = None
-    crypto_ed25519 = None
-    crypto_ed448 = None
-    crypto_hashes = None
-    crypto_padding = None
-    crypto_pkcs7 = None
-    crypto_rsa = None
-    crypto_serialization = None
-
-try:
-    import josepy
-except ImportError:
-    josepy = None
 
 # ------------------------------------------------------------------------------
 
@@ -129,7 +117,7 @@ if cryptography and josepy:
 
 # ==============================================================================
 
-log = logging.getLogger(__name__)
+log = logging.getLogger("cert_utils")
 log.setLevel(logging.INFO)
 
 # ------------------------------------------------------------------------------
@@ -147,7 +135,6 @@ openssl_path_conf = (
 ACME_VERSION = "v2"
 openssl_version: Optional[List[int]] = None
 _RE_openssl_version = re.compile(r"OpenSSL ((\d+\.\d+\.\d+)\w*) ", re.I)
-_RE_rn = re.compile(r"\r\n")
 _openssl_behavior: Optional[str] = None  # 'a' or 'b'
 
 
@@ -195,63 +182,6 @@ def update_from_appsettings(appsettings: Dict[str, Any]) -> None:
 # ==============================================================================
 
 
-# note the conditional whitespace before/after `CN`
-# this is because of differing openssl versions
-RE_openssl_x509_subject = re.compile(r"Subject:.*? CN ?= ?([^\s,;/]+)")
-RE_openssl_x509_san = re.compile(
-    r"X509v3 Subject Alternative Name: ?\n +([^\n]+)\n?", re.MULTILINE | re.DOTALL
-)
-
-
-# openssl 3 does not have "keyid:" as a prefix
-# a keyid prefix is okay!
-# we do not want the alternates, which are uri+serial; but take that out in results
-RE_openssl_x509_authority_key_identifier = re.compile(
-    r"X509v3 Authority Key Identifier: ?\n +(?:keyid:)?([^\n]+)\n?",
-    re.MULTILINE | re.DOTALL,
-)
-# we have a potential line in there for the OSCP or something else.
-RE_openssl_x509_issuer_uri = re.compile(
-    r"Authority Information Access: ?\n(?:[^\n]*^\n)? +CA Issuers - URI:([^\n]+)\n?",
-    re.MULTILINE | re.DOTALL,
-)
-
-RE_openssl_x509_serial = re.compile(r"Serial Number: ?(\d+)")
-
-#
-# https://github.com/certbot/certbot/blob/master/certbot/certbot/crypto_util.py#L482
-#
-# Finds one CERTIFICATE stricttextualmsg according to rfc7468#section-3.
-# Does not validate the base64text - use crypto.load_certificate.
-#
-# NOTE: this functions slightly differently as " *?" was added
-#       the first two letsencrypt certificates added a trailing space, which may
-#       not be compliant with the specification
-CERT_PEM_REGEX = re.compile(
-    """-----BEGIN CERTIFICATE----- *?\r?
-.+?\r?
------END CERTIFICATE----- *?\r?
-""",
-    re.DOTALL,  # DOTALL (/s) because the base64text may include newlines
-)
-
-# depending on openssl version, the "Public key: " text might list the bits
-# it may or may not also have a dash in the phrase "Public Key"
-# it may or may not be prefaced with the PublicKey type
-RE_openssl_x509_keytype_rsa = re.compile(
-    r"Subject Public Key Info:\n"
-    r"\s+Public Key Algorithm: rsaEncryption\n"
-    r"\s+(RSA )?Public(\ |\-)Key:",
-    re.MULTILINE,
-)
-RE_openssl_x509_keytype_ec = re.compile(
-    r"Subject Public Key Info:\n"
-    r"\s+Public Key Algorithm: id-ecPublicKey\n"
-    r"\s+(EC )?Public(\ |\-)Key:",
-    re.MULTILINE,
-)
-
-
 # see https://community.letsencrypt.org/t/issuing-for-common-rsa-key-sizes-only/133839
 # see https://letsencrypt.org/docs/integration-guide/
 ALLOWED_BITS_RSA = [2048, 3072, 4096]
@@ -282,314 +212,6 @@ EXTENSION_TO_MIME = {
         "*": "application/pkcs8",
     },
 }
-
-
-# ==============================================================================
-
-# General Utility Functions
-
-
-def new_pem_tempfile(pem_data: str) -> tempfile._TemporaryFileWrapper:
-    """
-    this is just a convenience wrapper to create a tempfile and seek(0)
-
-    :param pem_data: PEM encoded string to seed the tempfile with
-    :type pem_data: str
-    :returns: a tempfile instance
-    :rtype: tempfile.NamedTemporaryFile
-    """
-    tmpfile_pem = tempfile.NamedTemporaryFile()
-    if isinstance(pem_data, str):
-        pem_bytes = pem_data.encode()
-    tmpfile_pem.write(pem_bytes)
-    tmpfile_pem.seek(0)
-    return tmpfile_pem
-
-
-def new_der_tempfile(der_data: bytes) -> tempfile._TemporaryFileWrapper:
-    """
-    this is just a convenience wrapper to create a tempfile and seek(0)
-
-    :param der_data: DER encoded string to seed the tempfile with
-    :type der_data: str
-    :returns: a tempfile instance
-    :rtype: `tempfile.NamedTemporaryFile`
-    """
-    tmpfile_der = tempfile.NamedTemporaryFile()
-    tmpfile_der.write(der_data)
-    tmpfile_der.seek(0)
-    return tmpfile_der
-
-
-def cleanup_pem_text(pem_text: str) -> str:
-    """
-    * standardizes newlines;
-    * removes trailing spaces;
-    * ensures a trailing newline.
-
-    :param pem_text: PEM formatted string
-    :type pem_text: str
-    :returns: cleaned PEM text
-    :rtype: str
-    """
-    pem_text = _RE_rn.sub("\n", pem_text)
-    _pem_text_lines = [i.strip() for i in pem_text.split("\n")]
-    _pem_text_lines = [i for i in _pem_text_lines if i]
-    pem_text = "\n".join(_pem_text_lines) + "\n"
-    return pem_text
-
-
-def split_pem_chain(pem_text: str) -> List[str]:
-    """
-    splits a PEM encoded Certificate chain into multiple Certificates
-
-    :param pem_text: PEM formatted string containing one or more Certificates
-    :type pem_text: str
-    :returns: a list of PEM encoded Certificates
-    :rtype: list
-    """
-    _certs = CERT_PEM_REGEX.findall(pem_text)
-    certs = [cleanup_pem_text(i) for i in _certs]
-    return certs
-
-
-def convert_der_to_pem(der_data: bytes) -> str:
-    """
-    :param der_data: DER encoded string
-    :type der_data: str
-    :returns: PEM encoded version of the DER Certificate
-    :rtype: str
-    """
-    # PEM is just a b64 encoded DER Certificate with the header/footer
-    as_pem = """-----BEGIN CERTIFICATE-----\n{0}\n-----END CERTIFICATE-----\n""".format(
-        "\n".join(textwrap.wrap(base64.b64encode(der_data).decode("utf8"), 64))
-    )
-    return as_pem
-
-
-def convert_der_to_pem__csr(der_data: bytes) -> str:
-    """
-    :param der_data: CSR in DER encoding
-    :type der_data: str
-    :returns: PEM encoded version of the DER encoded CSR
-    :rtype: str
-    """
-    # PEM is just a b64 encoded DER CertificateRequest with the header/footer
-    as_pem = """-----BEGIN CERTIFICATE REQUEST-----\n{0}\n-----END CERTIFICATE REQUEST-----\n""".format(
-        "\n".join(textwrap.wrap(base64.b64encode(der_data).decode("utf8"), 64))
-    )
-    return as_pem
-
-
-def convert_der_to_pem__rsakey(der_data: bytes) -> str:
-    """
-    :param der_data: RSA Key in DER encoding
-    :type der_data: str
-    :returns: PEM encoded version of the RSA Key
-    :rtype: str
-
-    The OpenSSL Equivalent / Fallback is::
-
-        openssl rsa -in {FILEPATH} -inform der -outform pem
-    """
-    # PEM is just a b64 encoded DER RSA KEY with the header/footer
-    as_pem = """-----BEGIN RSA PRIVATE KEY-----\n{0}\n-----END RSA PRIVATE KEY-----\n""".format(
-        "\n".join(textwrap.wrap(base64.b64encode(der_data).decode("utf8"), 64))
-    )
-    return as_pem
-
-
-def convert_pem_to_der(pem_data: str) -> bytes:
-    """
-    :param pem_data: PEM encoded data
-    :type pem_data: str
-    :returns: DER encoded version of the PEM data
-    :rtype: str
-
-    The OpenSSL Equivalent / Fallback is::
-
-        openssl req -in {FILEPATH} -outform DER
-
-    The RFC requires the PEM header/footer to start/end with 5 dashes
-    This function is a bit lazy and does not check that.
-    """
-    # PEM is just a b64 encoded DER data with the appropiate header/footer
-    lines = [_l.strip() for _l in pem_data.strip().split("\n")]
-    # remove the BEGIN CERT
-    if (
-        ("BEGIN CERTIFICATE" in lines[0])
-        or ("BEGIN RSA PRIVATE KEY" in lines[0])
-        or ("BEGIN PRIVATE KEY" in lines[0])
-        or ("BEGIN CERTIFICATE REQUEST" in lines[0])
-    ):
-        lines = lines[1:]
-    if (
-        ("END CERTIFICATE" in lines[-1])
-        or ("END RSA PRIVATE KEY" in lines[-1])
-        or ("END PRIVATE KEY" in lines[-1])
-        or ("END CERTIFICATE REQUEST" in lines[-1])
-    ):
-        lines = lines[:-1]
-    stringed = "".join(lines)
-    result = base64.b64decode(stringed)
-    return result
-
-
-def convert_pkcs7_to_pems(pkcs7_data: bytes) -> List[str]:
-    """
-    :param pkcs7_data: pkcs7 encoded Certifcate Chain
-    :type pem_data: str
-    :returns: list of PEM encoded Certificates in the pkcs7_data
-    :rtype: list
-
-    The OpenSSL Equivalent / Fallback is::
-
-        openssl pkcs7 -inform DER -in {FILEPATH} -print_certs -outform PEM
-    """
-    # TODO: accept a pkcs7 filepath; FallbackError_FilepathRequired
-    log.info("convert_pkcs7_to_pems >")
-    if cryptography:
-        if TYPE_CHECKING:
-            assert crypto_pkcs7 is not None
-            assert crypto_serialization is not None
-        certs_loaded = crypto_pkcs7.load_der_pkcs7_certificates(pkcs7_data)
-        certs_bytes = [
-            cert.public_bytes(crypto_serialization.Encoding.PEM)
-            for cert in certs_loaded
-        ]
-        certs_string = [cert.decode("utf8") for cert in certs_bytes]
-        certs_string = [cleanup_pem_text(cert) for cert in certs_string]
-        return certs_string
-
-    log.debug(".convert_pkcs7_to_pems > openssl fallback")
-    if openssl_version is None:
-        check_openssl_version()
-
-    _tmpfile_der = new_der_tempfile(pkcs7_data)
-    try:
-        cert_der_filepath = _tmpfile_der.name
-        with psutil.Popen(
-            [
-                openssl_path,
-                "pkcs7",
-                "-inform",
-                "DER",
-                "-in",
-                cert_der_filepath,
-                "-print_certs",
-                "-outform",
-                "PEM",
-            ],
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-        ) as proc:
-            data_bytes, err = proc.communicate()
-            if not data_bytes:
-                raise OpenSslError_InvalidCertificate(err)
-            data_str = data_bytes.decode()
-            # OpenSSL might return extra info
-            # for example: "subject=/O=Digital Signature Trust Co./CN=DST Root CA X3\nissuer=/O=Digital Signature Trust Co./CN=DST Root CA X3\n-----BEGIN CERTIFICATE---[...]"
-            # split_pem_chain works perfectly with this payload!
-            certs = split_pem_chain(data_str)
-        return certs
-    except Exception as exc:  # noqa: F841
-        raise
-    finally:
-        _tmpfile_der.close()
-
-
-def san_domains_from_text(text: str) -> List[str]:
-    """
-    Helper function to extract SAN domains from a chunk of text in a x509 object
-
-    :param text: string extracted from a x509 document
-    :type text: str
-    :returns: list of domains
-    :rtype: list
-    """
-    san_domains = set([])
-    _subject_alt_names = RE_openssl_x509_san.search(text)
-    if _subject_alt_names is not None:
-        for _san in _subject_alt_names.group(1).split(", "):
-            if _san.startswith("DNS:"):
-                san_domains.add(_san[4:].lower())
-    return sorted(list(san_domains))
-
-
-def authority_key_identifier_from_text(text: str) -> Optional[str]:
-    """
-    :param text: string extracted from a x509 document
-    :type text: str
-    :returns: authority_key_identifier
-    :rtype: str
-
-    openssl will print a uppercase hex pairs, separated by a colon
-    we should remove the colons
-    """
-    results = RE_openssl_x509_authority_key_identifier.findall(text)
-    if results:
-        authority_key_identifier = results[0]
-        # ensure we have a key_id and not "URI:" or other convention
-        if authority_key_identifier[2] == ":":
-            return authority_key_identifier.replace(":", "")
-    return None
-
-
-def serial_from_text(text: str) -> Optional[int]:
-    """
-    :param text: string extracted from a x509 document
-    :type text: str
-    :returns: serial
-    :rtype: int
-    """
-    results = RE_openssl_x509_serial.findall(text)
-    if results:
-        serial = results[0]
-        return int(serial)
-    return None
-
-
-def issuer_uri_from_text(text: str) -> Optional[str]:
-    """
-    :param text: string extracted from a x509 document
-    :type text: str
-    :returns: issuer_uri
-    :rtype: str
-    """
-    results = RE_openssl_x509_issuer_uri.findall(text)
-    if results:
-        return results[0]
-    return None
-
-
-def _cert_pubkey_technology__text(cert_text: str) -> Optional[str]:
-    """
-    :param cert_text: string extracted from a x509 document
-    :type cert_text: str
-    :returns: Pubkey type: "RSA" or "EC"
-    :rtype: str
-    """
-    # `cert_text` is the output of of `openssl x509 -noout -text -in MYCERT `
-    if RE_openssl_x509_keytype_rsa.search(cert_text):
-        return "RSA"
-    elif RE_openssl_x509_keytype_ec.search(cert_text):
-        return "EC"
-    return None
-
-
-def _csr_pubkey_technology__text(csr_text: str) -> Optional[str]:
-    """
-    :param csr_text: string extracted from a CSR document
-    :type csr_text: str
-    :returns: Pubkey type: "RSA" or "EC"
-    :rtype: str
-    """
-    # `csr_text` is the output of of `openssl req -noout -text -in MYCERT`
-    if RE_openssl_x509_keytype_rsa.search(csr_text):
-        return "RSA"
-    elif RE_openssl_x509_keytype_ec.search(csr_text):
-        return "EC"
-    return None
 
 
 # ==============================================================================
@@ -639,427 +261,6 @@ def check_openssl_version(replace: bool = False) -> List[int]:
             # some regex are different, but the behavior should be the same
             _openssl_behavior = "b"
     return openssl_version
-
-
-def _openssl_cert__normalize_pem(cert_pem: str) -> str:
-    """
-    normalize a cert using openssl
-    NOTE: this is an openssl fallback routine
-
-    :param cert_pem: PEM encoded Certificate data
-    :type cert_pem: str
-    :returns: normalized Certificate
-    :rtype: str
-
-    This runs via OpenSSL:
-
-        openssl x509 -in {FILEPATH}
-    """
-    if openssl_version is None:
-        check_openssl_version()
-
-    _tmpfile_pem = new_pem_tempfile(cert_pem)
-    try:
-        cert_pem_filepath = _tmpfile_pem.name
-        with psutil.Popen(
-            [openssl_path, "x509", "-in", cert_pem_filepath],
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-        ) as proc:
-            data_bytes, err = proc.communicate()
-            if not data_bytes:
-                raise OpenSslError_InvalidCertificate(err)
-            data_str = data_bytes.decode("utf8")
-            data_str = data_str.strip()
-        return data_str
-    except Exception as exc:  # noqa: F841
-        raise
-    finally:
-        _tmpfile_pem.close()
-
-
-def _openssl_spki_hash_cert(
-    key_technology: str = "",
-    cert_pem_filepath: str = "",
-    as_b64: Optional[bool] = None,
-) -> str:
-    """
-    :param key_technology: Is the key an "EC" or "RSA" key?
-    :type key_technology: str
-    :param cert_pem_filepath: REQUIRED filepath to PEM Certificate.
-                              Used for commandline OpenSSL fallback operations.
-    :type cert_pem_filepath: str
-    :param as_b64: Should the result be returned in Base64 encoding? default None
-    :type as_b64: boolean
-    :returns: spki_hash
-    :rtype: str
-
-    The OpenSSL Equivalent / Fallback is::
-
-        openssl x509 -pubkey -noout -in {CERT_FILEPATH} | \
-        openssl {key_technology} -pubout -outform DER -pubin | \
-        openssl dgst -sha256 -binary | \
-        openssl enc -base64
-    """
-    if key_technology not in ("EC", "RSA"):
-        raise ValueError("must submit `key_technology`")
-    key_technology = key_technology.lower()
-    if not cert_pem_filepath:
-        raise FallbackError_FilepathRequired("Must submit `cert_pem_filepath`.")
-    if openssl_version is None:
-        check_openssl_version()
-
-    spki_hash = None
-    # convert to DER
-    p1 = p2 = p3 = proc4 = None
-    try:
-        # extract the key
-        p1 = psutil.Popen(
-            [
-                openssl_path,
-                "x509",
-                "-pubkey",
-                "-noout",
-                "-in",
-                cert_pem_filepath,
-            ],
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-        )
-        # convert to DER
-        p2 = psutil.Popen(
-            [
-                openssl_path,
-                key_technology,
-                "-pubin",
-                "-pubout",
-                "-outform",
-                "DER",
-            ],
-            stdin=p1.stdout,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-        )
-        # digest
-        p3 = psutil.Popen(
-            [
-                openssl_path,
-                "dgst",
-                "-sha256",
-                "-binary",
-            ],
-            stdin=p2.stdout,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-        )
-        # encode
-        spki_hash = None
-        if as_b64:
-            with psutil.Popen(
-                [
-                    openssl_path,
-                    "enc",
-                    "-base64",
-                ],
-                stdin=p3.stdout,
-                stdout=subprocess.PIPE,
-                stderr=subprocess.PIPE,
-            ) as proc4:
-                spki_hash, err = proc4.communicate()
-                if err:
-                    raise OpenSslError("could not generate SPKI Hash")
-        else:
-            spki_hash, err = p3.communicate()
-            if err:
-                raise OpenSslError("could not generate SPKI Hash")
-            spki_hash = binascii.b2a_hex(spki_hash)
-            spki_hash = spki_hash.upper()
-        spki_hash = spki_hash.strip()
-        spki_hash = spki_hash.decode("utf8")
-
-    finally:
-        # Note: explicitly close what we opened
-        for _p in (
-            p1,
-            p2,
-            p3,
-        ):
-            if _p is not None:
-                try:
-                    _p.stdout.close()
-                    _p.stderr.close()
-                    _p.terminate()
-                    _p.wait()
-                except psutil.NoSuchProcess:
-                    pass
-    return spki_hash
-
-
-def _openssl_spki_hash_csr(
-    key_technology: str = "",
-    csr_pem_filepath: str = "",
-    as_b64: Optional[bool] = None,
-) -> str:
-    """
-    :param key_technology: Is the key an "EC" or "RSA" key?
-    :type key_technology: str
-    :param csr_pem_filepath: REQUIRED filepath to PEM CSR.
-                             Used for commandline OpenSSL fallback operations.
-    :type csr_pem_filepath: str
-    :param as_b64: Should the result be returned in Base64 encoding? default None
-    :type as_b64: boolean
-    :returns: spki_hash
-    :rtype: str
-
-    The OpenSSL Equivalent / Fallback is::
-
-        openssl REQ -pubkey -noout -in {CSR_FILEPATH} | \
-        openssl {key_technology} -pubout -outform DER -pubin | \
-        openssl dgst -sha256 -binary | \
-        openssl enc -base64
-    """
-    if key_technology not in ("EC", "RSA"):
-        raise ValueError("must submit `key_technology`")
-    key_technology = key_technology.lower()
-    if not csr_pem_filepath:
-        raise FallbackError_FilepathRequired("Must submit `csr_pem_filepath`.")
-    if openssl_version is None:
-        check_openssl_version()
-
-    spki_hash = None
-    # convert to DER
-    p1 = p2 = p3 = proc4 = None
-    try:
-        # extract the key
-        p1 = psutil.Popen(
-            [
-                openssl_path,
-                "req",
-                "-pubkey",
-                "-noout",
-                "-in",
-                csr_pem_filepath,
-            ],
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-        )
-        # convert to DER
-        p2 = psutil.Popen(
-            [
-                openssl_path,
-                key_technology,
-                "-pubin",
-                "-pubout",
-                "-outform",
-                "DER",
-            ],
-            stdin=p1.stdout,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-        )
-        # digest
-        p3 = psutil.Popen(
-            [
-                openssl_path,
-                "dgst",
-                "-sha256",
-                "-binary",
-            ],
-            stdin=p2.stdout,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-        )
-        # encode
-        spki_hash = None
-        if as_b64:
-            with psutil.Popen(
-                [
-                    openssl_path,
-                    "enc",
-                    "-base64",
-                ],
-                stdin=p3.stdout,
-                stdout=subprocess.PIPE,
-                stderr=subprocess.PIPE,
-            ) as proc4:
-                spki_hash, err = proc4.communicate()
-                if err:
-                    raise OpenSslError("could not generate SPKI Hash")
-        else:
-            spki_hash, err = p3.communicate()
-            if err:
-                raise OpenSslError("could not generate SPKI Hash")
-            spki_hash = binascii.b2a_hex(spki_hash)
-            spki_hash = spki_hash.upper()
-        spki_hash = spki_hash.strip()
-        spki_hash = spki_hash.decode("utf8")
-    finally:
-        # Note: explicitly close what we opened
-        for _p in (
-            p1,
-            p2,
-            p3,
-        ):
-            if _p is not None:
-                try:
-                    _p.stdout.close()
-                    _p.stderr.close()
-                    _p.terminate()
-                    _p.wait()
-                except psutil.NoSuchProcess:
-                    pass
-    return spki_hash
-
-
-def _openssl_spki_hash_pkey(
-    key_technology: str = "",
-    key_pem_filepath: str = "",
-    as_b64: Optional[bool] = None,
-) -> str:
-    """
-    :param key_technology: Is the key an "EC" or "RSA" key?
-    :type key_technology: str
-    :param key_pem_filepath: REQUIRED filepath to PEM encoded PrivateKey.
-                             Used for commandline OpenSSL fallback operations.
-    :type key_pem_filepath: str
-    :param as_b64: Should the result be returned in Base64 encoding? default None
-    :type as_b64: boolean
-    :returns: spki_hash
-    :rtype: str
-
-    The OpenSSL Equivalent / Fallback is::
-
-        openssl rsa -in {KEY_FILEPATH} -pubout -outform der | \
-        openssl dgst -sha256 -binary | \
-        openssl enc -base64
-    """
-    if key_technology not in ("EC", "RSA"):
-        raise ValueError("must submit `key_technology`")
-    key_technology = key_technology.lower()
-    if not key_pem_filepath:
-        raise FallbackError_FilepathRequired("Must submit `key_pem_filepath`.")
-    if openssl_version is None:
-        check_openssl_version()
-
-    spki_hash = None
-    # convert to DER
-    p1 = p2 = proc3 = None
-    try:
-        # convert to DER
-        p1 = psutil.Popen(
-            [
-                openssl_path,
-                key_technology,
-                "-pubout",
-                "-outform",
-                "DER",
-                "-in",
-                key_pem_filepath,
-            ],
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-        )
-        # digest
-        p2 = psutil.Popen(
-            [
-                openssl_path,
-                "dgst",
-                "-sha256",
-                "-binary",
-            ],
-            stdin=p1.stdout,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-        )
-        # encode
-        if as_b64:
-            with psutil.Popen(
-                [
-                    openssl_path,
-                    "enc",
-                    "-base64",
-                ],
-                stdin=p2.stdout,
-                stdout=subprocess.PIPE,
-                stderr=subprocess.PIPE,
-            ) as proc3:
-                spki_hash, err = proc3.communicate()
-                if err:
-                    raise OpenSslError("could not generate SPKI Hash")
-        else:
-            spki_hash, err = p2.communicate()
-            if err:
-                raise OpenSslError("could not generate SPKI Hash")
-            spki_hash = binascii.b2a_hex(spki_hash)
-            spki_hash = spki_hash.upper()
-        spki_hash = spki_hash.strip()
-        spki_hash = spki_hash.decode("utf8")
-    finally:
-        # Note: explicitly close what we opened
-        for _p in (
-            p1,
-            p2,
-        ):
-            if _p is not None:
-                try:
-                    _p.stdout.close()
-                    _p.stderr.close()
-                    _p.terminate()
-                    _p.wait()
-                except psutil.NoSuchProcess:
-                    pass
-    return spki_hash
-
-
-# ==============================================================================
-
-
-def _cryptography__public_key_technology(
-    publickey: "_TYPES_CRYPTOGRAPHY_PUBLICKEY",
-) -> Optional[str]:
-    """
-    :param key: key object
-    :type key: instance of
-    :returns: type of key: RSA, EC, DSA
-    :rtype: str
-    """
-    if TYPE_CHECKING:
-        assert crypto_rsa is not None
-        assert crypto_ec is not None
-    if isinstance(publickey, crypto_rsa.RSAPublicKey):
-        return "RSA"
-    elif isinstance(publickey, crypto_ec.EllipticCurvePublicKey):
-        return "EC"
-    return None
-
-
-def _cryptography__public_key_spki_sha256(
-    cryptography_publickey: Union[
-        "_TYPES_CRYPTOGRAPHY_KEYS", "_TYPES_CRYPTOGRAPHY_PUBLICKEY_EXTENDED"
-    ],
-    as_b64: Optional[bool] = None,
-) -> str:
-    """
-    :param cryptography_publickey: a PublicKey from the cryptography package
-    :type cryptography_publickey: cryptography.hazmat.backends.openssl.rsa._RSAPublicKey
-    :param as_b64: Should the result be returned in Base64 encoding? default None
-    :type as_b64: boolean
-    :returns: spki_sha256
-    :rtype: str
-    """
-    assert crypto_serialization is not None  # nest under `if TYPE_CHECKING` not needed
-    _public_bytes = cryptography_publickey.public_bytes(  # type: ignore[union-attr]
-        crypto_serialization.Encoding.DER,
-        crypto_serialization.PublicFormat.SubjectPublicKeyInfo,
-    )
-    spki_sha256 = hashlib.sha256(_public_bytes).digest()
-    if as_b64:
-        spki_sha256 = base64.b64encode(spki_sha256)
-    else:
-        spki_sha256 = binascii.b2a_hex(spki_sha256)
-        spki_sha256 = spki_sha256.upper()
-    _spki_sha256 = spki_sha256.decode("utf8")
-    return _spki_sha256
 
 
 # ==============================================================================
@@ -1299,33 +500,6 @@ def make_csr(
         raise OpenSslError_CsrGeneration("invalid ACME generator")
 
     return csr_text
-
-
-def _cryptography__cert_or_req__all_names(
-    loaded_cert_or_req: Union["Certificate", "CertificateSigningRequest"],
-) -> List[str]:
-    assert cryptography is not None
-    cn: Optional[str] = None
-    try:
-        _cn = loaded_cert_or_req.subject.get_attributes_for_oid(
-            cryptography.x509.oid.NameOID.COMMON_NAME
-        )[0].value
-        cn = _cn.decode() if isinstance(_cn, bytes) else _cn
-    except IndexError:
-        pass
-    sans: List[str] = []
-    try:
-        ext = loaded_cert_or_req.extensions.get_extension_for_oid(
-            cryptography.x509.oid.ExtensionOID.SUBJECT_ALTERNATIVE_NAME
-        )
-        _sans = ext.value.get_values_for_type(cryptography.x509.DNSName)  # type: ignore[attr-defined]
-        sans = [san.decode() if isinstance(san, bytes) else san for san in _sans]
-
-    except cryptography.x509.ExtensionNotFound:
-        pass
-    if cn is None:
-        return sans
-    return [cn] + [d for d in sans if d != cn]
 
 
 def parse_cert__domains(
@@ -1766,135 +940,6 @@ def fingerprint_cert(
     return data_str
 
 
-def _cleanup_openssl_md5(data: bytes) -> str:
-    """
-    some versions of openssl handle the md5 as:
-        '1231231231'
-    others handle as
-        "(stdin)= 123123'
-    """
-    data = data.strip()
-    data_str = data.decode("utf8")
-    if len(data_str) == 32 and (data_str[:9] != "(stdin)= "):
-        return data_str
-    if data_str[:9] != "(stdin)= " or not data_str:
-        raise OpenSslError("error reading md5 (i)")
-    data_str = data_str[9:]
-    if len(data_str) != 32:
-        raise OpenSslError("error reading md5 (ii)")
-    return data_str
-
-
-def _cleanup_openssl_modulus(data: str) -> str:
-    data = data.strip()
-    if data[:8] == "Modulus=":
-        data = data[8:]
-    return data
-
-
-def _format_crypto_components(
-    data: Union[
-        List[str],
-        List[Tuple[str, ...]],
-        List[Tuple[bytes, bytes]],
-    ],
-    fieldset: Optional[str] = None,
-) -> str:
-    """
-    :param data: input
-    :param fieldset: is unused. would be "issuer" or "subject"
-
-    `get_components()` is somewhat structured
-    the following are valid:
-    * [('CN', 'Pebble Intermediate CA 601ea1')]
-    * [('C', 'US'), ('O', 'Internet Security Research Group'), ('CN', 'ISRG Root X2')]
-    * [('C', 'US'), ('O', 'Internet Security Research Group'), ('CN', 'ISRG Root X1')]
-    * [('O', 'Digital Signature Trust Co.'), ('CN', 'DST Root CA X3')]
-    cert = openssl_crypto.load_certificate(openssl_crypto.FILETYPE_PEM, cert_pem)
-    _issuer = cert.get_issuer().get_components()
-    _subject = cert.get_subject().get_components()
-    """
-    _out = []
-    for _in_set in data:
-        _converted = [i.decode("utf8") if isinstance(i, bytes) else i for i in _in_set]  # type: ignore[attr-defined]
-        _out.append("=".join(_converted))
-    out = "\n".join(_out).strip()
-    return out
-
-
-def _format_cryptography_components(
-    x509_name: "Name",
-) -> str:
-    """
-    :param data: input
-
-
-    `get_components()` is somewhat structured
-    the following are valid:
-    * [('CN', 'Pebble Intermediate CA 601ea1')]
-    * [('C', 'US'), ('O', 'Internet Security Research Group'), ('CN', 'ISRG Root X2')]
-    * [('C', 'US'), ('O', 'Internet Security Research Group'), ('CN', 'ISRG Root X1')]
-    * [('O', 'Digital Signature Trust Co.'), ('CN', 'DST Root CA X3')]
-    cert = openssl_crypto.load_certificate(openssl_crypto.FILETYPE_PEM, cert_pem)
-    _issuer = cert.get_issuer().get_components()
-    _subject = cert.get_subject().get_components()
-    """
-    # ";".join(["%s=%s" % (i[0], i[1]) for i in x509_name])
-    _out = []
-    for attr in x509_name:
-        _converted = [i.decode("utf8") if isinstance(i, bytes) else i for i in (attr.rfc4514_attribute_name, attr.value)]  # type: ignore[attr-defined]
-        _out.append("=".join(_converted))
-    out = "\n".join(_out).strip()
-    return out
-
-
-def _format_openssl_components(
-    data: str,
-    fieldset: Optional[str] = None,
-) -> str:
-    """
-    different openssl versions give different responses. FUN.
-
-    To make things easier, just format this into the crypto compatible payload,
-    then invoke the crypto formattter
-
-    openssl = [0, 9, 8]
-    subject= /C=US/O=Internet Security Research Group/CN=ISRG Root X2
-
-    openssl = [1, 1, 1]
-    issuer=C = US, O = Internet Security Research Group, CN = ISRG Root X2
-    """
-    # print(openssl_version, data)
-    if fieldset in ("issuer", "subject"):
-        if fieldset == "issuer":
-            if data.startswith("issuer= "):
-                data = data[8:]
-            elif data.startswith("issuer="):
-                data = data[7:]
-        elif fieldset == "subject":
-            if data.startswith("subject= "):
-                data = data[9:]
-            elif data.startswith("subject="):
-                data = data[8:]
-        data_list: List[str]
-        if "/" in data:
-            data_list = [i.strip() for i in data.split("/")]
-        elif "," in data:
-            data_list = [i.strip() for i in data.split(",")]
-        else:
-            data_list = [
-                data,
-            ]
-        _out = []
-        for _cset in data_list:
-            _cset_split = _cset.split("=")
-            _cset_edited = tuple(i.strip() for i in _cset_split)
-            _out.append(_cset_edited)
-        return _format_crypto_components(_out, fieldset=fieldset)
-    else:
-        raise ValueError("invalid fieldset")
-
-
 def modulus_md5_key(
     key_pem: str,
     key_pem_filepath: Optional[str] = None,
@@ -2073,231 +1118,6 @@ def modulus_md5_cert(
                 return None
     data_bytes = data_str.encode()
     data_str = hashlib.md5(data_bytes).hexdigest()
-    return data_str
-
-
-def _openssl_cert_single_op__pem(
-    cert_pem: str,
-    single_op: str,
-) -> str:
-    """
-    this just invokes `_openssl_cert_single_op__pem_filepath` with a tempfile
-    """
-    _tmpfile_pem = new_pem_tempfile(cert_pem)
-    try:
-        cert_pem_filepath = _tmpfile_pem.name
-        return _openssl_cert_single_op__pem_filepath(cert_pem_filepath, single_op)
-    except Exception as exc:  # noqa: F841
-        raise
-    finally:
-        _tmpfile_pem.close()
-
-
-def _openssl_cert_single_op__pem_filepath(
-    pem_filepath: str,
-    single_op: str,
-) -> str:
-    """
-    handles a single pem operation to `openssl x509`
-
-    :param pem_filepath: filepath to pem encoded cert
-    :type pem_filepath: str
-    :param single_op: operation
-    :type single_op: str
-    :returns: openssl output
-    :rtype: str
-
-    openssl x509 -noout -issuer -in cert.pem
-    openssl x509 -noout -issuer_hash -in cert.pem
-
-    openssl x509 -noout -issuer_hash -in {CERT}
-    returns the data found in
-       X509v3 extensions:
-           X509v3 Authority Key Identifier:
-               keyid:{VALUE}
-
-    openssl x509 -noout -subject_hash -in {CERT}
-    returns the data found in
-       X509v3 extensions:
-           X509v3 Subject Key Identifier:
-               {VALUE}
-
-    The OpenSSL Equivalent / Fallback is::
-
-        openssl x509 -noout {OPERATION} -in {FILEPATH})
-    """
-    if single_op not in (
-        "-issuer_hash",
-        "-issuer",
-        "-subject_hash",
-        "-subject",
-        "-startdate",
-        "-enddate",
-    ):
-        raise ValueError("invalid `single_op`")
-    if not pem_filepath:
-        raise FallbackError_FilepathRequired("Must submit `pem_filepath`.")
-    if openssl_version is None:
-        check_openssl_version()
-
-    with psutil.Popen(
-        [openssl_path, "x509", "-noout", single_op, "-in", pem_filepath],
-        stdout=subprocess.PIPE,
-        stderr=subprocess.PIPE,
-    ) as proc:
-        data_bytes, err = proc.communicate()
-        if not data_bytes:
-            raise OpenSslError_InvalidCertificate(err)
-        data_str = data_bytes.decode("utf8")
-        data_str = data_str.strip()
-    return data_str
-
-
-def cert_ext__pem_filepath(
-    pem_filepath: str,
-    ext: str,
-) -> str:
-    """
-    handles a single pem operation to `openssl x509` with EXTENSION
-    /usr/local/bin/openssl x509  -noout -ext subjectAltName -in cert.pem
-    /usr/local/bin/openssl x509  -noout -ext authorityKeyIdentifier -in cert.pem
-    /usr/local/bin/openssl x509  -noout -ext authorityInfoAccess -in cert.pem
-
-    :param pem_filepath: filepath to the PEM encoded Certificate
-    :type pem_filepath: str
-    :param ext: a supported x509 extension
-    :type ext: str
-    :returns: openssl output value
-    :rtype: str
-
-    The OpenSSL Equivalent / Fallback is::
-
-        openssl x509 -noout -ext {EXT} -in {FILEPATH})
-    """
-    if ext not in ("subjectAltName", "authorityKeyIdentifier", "authorityInfoAccess"):
-        raise ValueError("invalid `ext`")
-    if not pem_filepath:
-        raise FallbackError_FilepathRequired("Must submit `pem_filepath`.")
-    if openssl_version is None:
-        check_openssl_version()
-
-    with psutil.Popen(
-        [openssl_path, "x509", "-noout", "-ext", ext, "-in", pem_filepath],
-        stdout=subprocess.PIPE,
-        stderr=subprocess.PIPE,
-    ) as proc:
-        data_bytes, err = proc.communicate()
-        if not data_bytes:
-            raise OpenSslError_InvalidCertificate(err)
-        data_str = data_bytes.decode("utf8")
-        data_str = data_str.strip()
-    return data_str
-
-
-def csr_single_op__pem_filepath(
-    pem_filepath: str,
-    single_op: str,
-) -> str:
-    """
-    handles a single pem operation to `openssl req` with EXTENSION
-
-    openssl req -noout -subject -in csr.pem
-
-    :param pem_filepath: filepath to the PEM encoded CSR.
-    :type pem_filepath: str
-    :param single_op: a supported `openssl req` operation
-    :type single_op: str
-    :returns: openssl output value
-    :rtype: str
-
-    The OpenSSL Equivalent / Fallback is::
-
-        openssl req -noout {OPERATION} -in {FILEPATH})
-    """
-    if single_op not in ("-subject",):
-        raise ValueError("invalid `single_op`")
-    if not pem_filepath:
-        raise FallbackError_FilepathRequired("Must submit `pem_filepath`.")
-    if openssl_version is None:
-        check_openssl_version()
-
-    with psutil.Popen(
-        [openssl_path, "req", "-noout", single_op, "-in", pem_filepath],
-        stdout=subprocess.PIPE,
-        stderr=subprocess.PIPE,
-    ) as proc:
-        data_bytes, err = proc.communicate()
-        if not data_bytes:
-            raise OpenSslError_InvalidCSR(err)
-        data_str = data_bytes.decode("utf8")
-        data_str = data_str.strip()
-    return data_str
-
-
-def key_single_op__pem_filepath(
-    keytype: str = "RSA",
-    pem_filepath: str = "",
-    single_op: str = "",
-) -> str:
-    """
-    :param keytype: the type of key: RSA or EC
-    :type keytype: str
-    :param pem_filepath: filepath to the PEM encoded Key
-    :type pem_filepath: str
-    :param single_op: a supported `openssl rsa/ec` operation
-    :type single_op: str
-    :returns: openssl output value
-    :rtype: str
-
-    THIS SHOULD NOT BE USED BY INTERNAL CODE
-
-    This is a bit odd...
-
-    1. If "-check" is okay (or reading is okay), there may be no output on stdout
-       HOWEVER
-       the read message (success) may happen on stderr
-    2. If openssl can't read the file, it will raise an exception
-
-    earlier versions of openssl DO NOT HAVE `ec --check`
-    current versions do
-
-    The OpenSSL Equivalent / Fallback is::
-
-        openssl {KEYTYPE} -noout {OPERATION} -in {FILEPATH})
-
-    Such as:
-
-        openssl rsa -noout -check -in {KEY}
-        openssl rsa -noout -modulus -in {KEY}
-        openssl rsa -noout -text -in {KEY}
-
-        openssl ec -noout -in {KEY}
-        openssl ec -noout -modulus -in {KEY}
-        openssl ec -noout -text -in {KEY}
-    """
-    if keytype not in ("RSA", "EC"):
-        raise ValueError("keytype must be `RSA or EC`")
-    if single_op not in ("-check", "-modulus", "-text"):
-        raise ValueError("invalid `single_op`")
-    if not pem_filepath:
-        raise FallbackError_FilepathRequired("Must submit `pem_filepath`.")
-    if openssl_version is None:
-        check_openssl_version()
-
-    with psutil.Popen(
-        [openssl_path, keytype.lower(), "-noout", single_op, "-in", pem_filepath],
-        stdout=subprocess.PIPE,
-        stderr=subprocess.PIPE,
-    ) as proc:
-        data_bytes, err = proc.communicate()
-        if not data_bytes:
-            if err.startswith(b"unknown option -check"):
-                raise OpenSslError_VersionTooLow(err)
-            elif err != b"read EC key\nEC Key valid.\n":
-                # this happens, where some versions give an error and no data!
-                raise OpenSslError_InvalidKey(err)
-        data_str = data_bytes.decode("utf8")
-        data_str = data_str.strip()
     return data_str
 
 
@@ -3262,137 +2082,6 @@ def new_key_rsa(bits: int = 4096) -> str:
     return key_pem_str
 
 
-def convert_jwk_to_ans1(pkey_jsons: str) -> str:
-    """
-    input is a json string
-
-    adapted from https://github.com/JonLundy
-    who shared this gist under the MIT license:
-        https://gist.github.com/JonLundy/f25c99ee0770e19dc595
-
-    :param pkey_jsons: JWK Key
-    :type pkey_jsons: str
-    :returns: Key in ANS1 Format
-    :rtype: str
-    """
-    pkey = json.loads(pkey_jsons)
-
-    def enc(data_bytes: bytes) -> str:
-        missing_padding = 4 - len(data_bytes) % 4
-        if missing_padding:
-            data_bytes += b"=" * missing_padding
-        data_bytes = binascii.hexlify(base64.b64decode(data_bytes, b"-_")).upper()
-        data_str = data_bytes.decode("utf8")
-        return "0x" + data_str
-
-    for k, v in list(pkey.items()):
-        if k == "kty":
-            continue
-        pkey[k] = enc(v.encode())
-
-    converted = []
-    converted.append("asn1=SEQUENCE:private_key\n[private_key]\nversion=INTEGER:0")
-    converted.append("n=INTEGER:{}".format(pkey["n"]))
-    converted.append("e=INTEGER:{}".format(pkey["e"]))
-    converted.append("d=INTEGER:{}".format(pkey["d"]))
-    converted.append("p=INTEGER:{}".format(pkey["p"]))
-    converted.append("q=INTEGER:{}".format(pkey["q"]))
-    converted.append("dp=INTEGER:{}".format(pkey["dp"]))
-    converted.append("dq=INTEGER:{}".format(pkey["dq"]))
-    converted.append("qi=INTEGER:{}".format(pkey["qi"]))
-    converted.append("")  # trailing newline
-    converted_ = "\n".join(converted)
-    return converted_
-
-
-def convert_lejson_to_pem(pkey_jsons: str) -> str:
-    """
-    This routine will use crypto/certbot if available.
-    If not, openssl is used via subprocesses
-
-    input is a json string
-
-    adapted from https://github.com/JonLundy
-    who shared this gist under the MIT license:
-        https://gist.github.com/JonLundy/f25c99ee0770e19dc595
-
-    openssl asn1parse -noout -out private_key.der -genconf <(python conv.py private_key.json)
-    openssl rsa -in private_key.der -inform der > private_key.pem
-    openssl rsa -in private_key.pem
-
-    :param pkey_jsons: LetsEncrypt JSON formatted Key
-    :type pkey_jsons: str
-    :returns: Key in PEM Encoding
-    :rtype: str
-    """
-    log.info("convert_lejson_to_pem >")
-
-    if cryptography and josepy:
-        if TYPE_CHECKING:
-            assert crypto_serialization is not None
-        pkey = josepy.JWKRSA.json_loads(pkey_jsons)
-        as_pem = pkey.key.private_bytes(
-            encoding=crypto_serialization.Encoding.PEM,
-            format=crypto_serialization.PrivateFormat.TraditionalOpenSSL,
-            encryption_algorithm=crypto_serialization.NoEncryption(),
-        )
-        as_pem = as_pem.decode("utf8")
-        as_pem = cleanup_pem_text(as_pem)
-
-        # note: we don't need to provide key_pem_filepath because we already rely on openssl
-        key_technology = validate_key(key_pem=as_pem)
-        return as_pem
-
-    log.debug(".convert_lejson_to_pem > openssl fallback")
-    if openssl_version is None:
-        check_openssl_version()
-
-    pkey_ans1 = convert_jwk_to_ans1(pkey_jsons)
-    as_pem = None
-    tmpfiles = []
-    try:
-        tmpfile_ans1 = new_pem_tempfile(pkey_ans1)
-        tmpfiles.append(tmpfile_ans1)
-
-        tmpfile_der = new_pem_tempfile("")
-        tmpfiles.append(tmpfile_der)
-
-        with psutil.Popen(
-            [
-                openssl_path,
-                "asn1parse",
-                "-noout",
-                "-out",
-                tmpfile_der.name,
-                "-genconf",
-                tmpfile_ans1.name,
-            ],
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-        ) as proc:
-            generated, err = proc.communicate()
-            if err:
-                raise ValueError(err)
-        # convert to pem
-        as_pem = convert_der_to_pem__rsakey(tmpfile_der.read())
-
-        # we need a tmpfile to validate it
-        tmpfile_pem = new_pem_tempfile(as_pem)
-        tmpfiles.append(tmpfile_pem)
-
-        # validate it
-        key_technology = validate_key(  # noqa: F841
-            key_pem=as_pem, key_pem_filepath=tmpfile_pem.name
-        )
-        return as_pem
-
-    except Exception as exc:  # noqa: F841
-        raise
-    finally:
-        for t in tmpfiles:
-            t.close()
-
-
 def cert_and_chain_from_fullchain(
     fullchain_pem: str,
 ) -> Tuple[str, str]:
@@ -3416,6 +2105,7 @@ def cert_and_chain_from_fullchain(
         try:
             return cryptography__cert_and_chain_from_fullchain(fullchain_pem)
         except Exception as exc:
+            raise
             raise OpenSslError(exc)
 
     log.debug(".cert_and_chain_from_fullchain > openssl fallback")
@@ -3738,43 +2428,6 @@ def ensure_chain_order(
 
 
 # ------------------------------------------------------------------------------
-
-
-def jose_b64(b: bytes) -> str:
-    # helper function base64 encode for jose spec
-    return base64.urlsafe_b64encode(b).decode("utf8").replace("=", "")
-
-
-class AccountKeyData(object):
-    """
-    An object encapsulating Account Key data
-    """
-
-    key_pem: str
-    key_pem_filepath: Optional[str]
-    jwk: Dict
-    thumbprint: str
-    alg: str
-
-    def __init__(
-        self,
-        key_pem: str,
-        key_pem_filepath: Optional[str] = None,
-    ):
-        """
-        :param key_pem: (required) A PEM encoded RSA key
-        :param key_pem_filepath: (optional) The filepath of a PEM encoded RSA key
-        """
-        self.key_pem = key_pem
-        self.key_pem_filepath = key_pem_filepath
-
-        (_jwk, _thumbprint, _alg) = account_key__parse(
-            key_pem=key_pem,
-            key_pem_filepath=key_pem_filepath,
-        )
-        self.jwk = _jwk
-        self.thumbprint = _thumbprint
-        self.alg = _alg
 
 
 def account_key__parse(
